# go/api/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    REST API for managing, executing, and monitoring workflows, human tasks, incidents, and outgoing webhooks inside the NativeBPM Cloud-Native engine.

    ### Authentication
    Requests must include a session cookie or a Bearer API token:
    `Authorization: Bearer <API_TOKEN>`
  title: NativeBPM Engine REST API
  version: 1.0.0
servers:
- url: /
paths:
  /api/definitions:
    get:
      description: Retrieve a list of all deployed process definitions.
      operationId: listDefinitions
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ProcessDefinition"
                type: array
          description: Successful response
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listDefinitions_401_response"
          description: Unauthorized - missing or invalid session cookie / API Bearer
            Token
        "500":
          description: Internal Server Error - database failure or execution crash
      summary: List process definitions
  /api/deploy:
    post:
      description: Deploy a new BPMN 2.0 XML process definition.
      operationId: deployDefinition
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/deployDefinition_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessDefinition"
          description: Successful response
        "400":
          description: Invalid request body or malformed BPMN XML
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listDefinitions_401_response"
          description: Unauthorized - missing or invalid session cookie / API Bearer
            Token
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listDefinitions_401_response"
          description: Forbidden - insufficient operator permissions or role level
        "500":
          description: Internal Server Error - database failure or execution crash
      summary: Deploy process definition
  /api/definitions/{id}/start:
    post:
      description: Start a new workflow process instance by process definition ID.
      operationId: startInstance
      parameters:
      - description: "The process definition ID (e.g., matching the BPMN process element\
          \ ID)"
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartInstanceRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessInstance"
          description: Successful response
        "400":
          description: Invalid variables or configuration
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listDefinitions_401_response"
          description: Unauthorized - missing or invalid session cookie / API Bearer
            Token
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listDefinitions_401_response"
          description: Forbidden - insufficient operator permissions or role level
        "500":
          description: Internal Server Error - database failure or execution crash
      summary: Start process instance
...
components:
...
  schemas:
    ProcessDefinition:
      example:
        hash: hash
        id: id
        name: name
        xml_data: xml_data
        deployed_at: 2000-01-23T04:56:07.000+00:00
      properties:
        hash:
          description: MD5/SHA256 content hash of the process XML schema definition
          type: string
        id:
          description: Unique process definition identifier
          type: string
        name:
          description: Friendly name of the process
          type: string
        xml_data:
          description: Base64-encoded raw BPMN 2.0 XML schema data
          format: byte
          type: string
        deployed_at:
          format: date-time
          type: string
      required:
      - deployed_at
      - hash
      - id
      - name
      - xml_data
      type: object
    ProcessInstance:
      example:
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        process_id: process_id
        definition_hash: definition_hash
        business_key: business_key
        state: state
        version: 0
        completed: true
        updated_at: 2000-01-23T04:56:07.000+00:00
        tenant_id: tenant_id
      properties:
        id:
          format: uuid
          type: string
        process_id:
          type: string
        definition_hash:
          type: string
        business_key:
          type: string
        state:
          description: Base64/raw JSON encoded internal Wazero process engine state
            representation
          format: byte
          type: string
        version:
          type: integer
        completed:
          type: boolean
        updated_at:
          format: date-time
          type: string
        tenant_id:
          type: string
      required:
      - business_key
      - completed
      - definition_hash
      - id
      - process_id
      - state
      - tenant_id
      - updated_at
      - version
      type: object
...
    StartInstanceRequest:
      example:
        instance_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        business_key: business_key
        variables:
          key: ""
      properties:
        instance_id:
          description: Optional user-generated UUID to enforce idempotency
          format: uuid
          type: string
        business_key:
          description: Business tracking keyword
          type: string
        variables:
          additionalProperties: true
          type: object
      type: object
...
    listDefinitions_401_response:
      example:
        error: Unauthorized
      properties:
        error:
          example: Unauthorized
          type: string
      type: object
    deployDefinition_request:
      properties:
        file:
          description: BPMN 2.0 XML file content to deploy
          format: binary
          type: string
      type: object
...