# go/api/model_process_definition.go
/*
NativeBPM Engine REST API

REST API for managing, executing, and monitoring workflows, human tasks, incidents, and outgoing webhooks inside the NativeBPM Cloud-Native engine.  ### Authentication Requests must include a session cookie or a Bearer API token: `Authorization: Bearer <API_TOKEN>` 

API version: 1.0.0
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package api

import (
	"encoding/json"
	"time"
	"bytes"
	"fmt"
)

// checks if the ProcessDefinition type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &ProcessDefinition{}

// ProcessDefinition struct for ProcessDefinition
type ProcessDefinition struct {
	// MD5/SHA256 content hash of the process XML schema definition
	Hash string `json:"hash"`
	// Unique process definition identifier
	Id string `json:"id"`
	// Friendly name of the process
	Name string `json:"name"`
	// Base64-encoded raw BPMN 2.0 XML schema data
	XmlData string `json:"xml_data"`
	DeployedAt time.Time `json:"deployed_at"`
}

type _ProcessDefinition ProcessDefinition

// NewProcessDefinition instantiates a new ProcessDefinition object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewProcessDefinition(hash string, id string, name string, xmlData string, deployedAt time.Time) *ProcessDefinition {
	this := ProcessDefinition{}
	this.Hash = hash
	this.Id = id
	this.Name = name
	this.XmlData = xmlData
	this.DeployedAt = deployedAt
	return &this
}

// NewProcessDefinitionWithDefaults instantiates a new ProcessDefinition object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewProcessDefinitionWithDefaults() *ProcessDefinition {
	this := ProcessDefinition{}
	return &this
}

// GetHash returns the Hash field value
func (o *ProcessDefinition) GetHash() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.Hash
}

// GetHashOk returns a tuple with the Hash field value
// and a boolean to check if the value has been set.
func (o *ProcessDefinition) GetHashOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Hash, true
}

// SetHash sets field value
func (o *ProcessDefinition) SetHash(v string) {
	o.Hash = v
}

// GetId returns the Id field value
func (o *ProcessDefinition) GetId() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.Id
}

// GetIdOk returns a tuple with the Id field value
// and a boolean to check if the value has been set.
func (o *ProcessDefinition) GetIdOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Id, true
}

// SetId sets field value
func (o *ProcessDefinition) SetId(v string) {
	o.Id = v
}

// GetName returns the Name field value
func (o *ProcessDefinition) GetName() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.Name
}

// GetNameOk returns a tuple with the Name field value
// and a boolean to check if the value has been set.
func (o *ProcessDefinition) GetNameOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Name, true
}

// SetName sets field value
func (o *ProcessDefinition) SetName(v string) {
	o.Name = v
}

// GetXmlData returns the XmlData field value
func (o *ProcessDefinition) GetXmlData() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.XmlData
}

// GetXmlDataOk returns a tuple with the XmlData field value
// and a boolean to check if the value has been set.
func (o *ProcessDefinition) GetXmlDataOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.XmlData, true
}

// SetXmlData sets field value
func (o *ProcessDefinition) SetXmlData(v string) {
	o.XmlData = v
}

// GetDeployedAt returns the DeployedAt field value
func (o *ProcessDefinition) GetDeployedAt() time.Time {
	if o == nil {
		var ret time.Time
		return ret
	}

	return o.DeployedAt
}

// GetDeployedAtOk returns a tuple with the DeployedAt field value
// and a boolean to check if the value has been set.
func (o *ProcessDefinition) GetDeployedAtOk() (*time.Time, bool) {
	if o == nil {
		return nil, false
	}
	return &o.DeployedAt, true
}

// SetDeployedAt sets field value
func (o *ProcessDefinition) SetDeployedAt(v time.Time) {
	o.DeployedAt = v
}

func (o ProcessDefinition) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o ProcessDefinition) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["hash"] = o.Hash
	toSerialize["id"] = o.Id
	toSerialize["name"] = o.Name
	toSerialize["xml_data"] = o.XmlData
	toSerialize["deployed_at"] = o.DeployedAt
	return toSerialize, nil
}

func (o *ProcessDefinition) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"hash",
		"id",
		"name",
		"xml_data",
		"deployed_at",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err;
	}

	for _, requiredProperty := range(requiredProperties) {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varProcessDefinition := _ProcessDefinition{}

	decoder := json.NewDecoder(bytes.NewReader(data))
	decoder.DisallowUnknownFields()
	err = decoder.Decode(&varProcessDefinition)

	if err != nil {
		return err
	}

	*o = ProcessDefinition(varProcessDefinition)

	return err
}

type NullableProcessDefinition struct {
	value *ProcessDefinition
	isSet bool
}

func (v NullableProcessDefinition) Get() *ProcessDefinition {
	return v.value
}

func (v *NullableProcessDefinition) Set(val *ProcessDefinition) {
	v.value = val
	v.isSet = true
}

func (v NullableProcessDefinition) IsSet() bool {
	return v.isSet
}

func (v *NullableProcessDefinition) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableProcessDefinition(val *ProcessDefinition) *NullableProcessDefinition {
	return &NullableProcessDefinition{value: val, isSet: true}
}

func (v NullableProcessDefinition) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableProcessDefinition) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}