# go/api/model_start_instance_request.go
/*
NativeBPM Engine REST API

REST API for managing, executing, and monitoring workflows, human tasks, incidents, and outgoing webhooks inside the NativeBPM Cloud-Native engine.  ### Authentication Requests must include a session cookie or a Bearer API token: `Authorization: Bearer <API_TOKEN>` 

API version: 1.0.0
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package api

import (
	"encoding/json"
)

// checks if the StartInstanceRequest type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &StartInstanceRequest{}

// StartInstanceRequest struct for StartInstanceRequest
type StartInstanceRequest struct {
	// Optional user-generated UUID to enforce idempotency
	InstanceId *string `json:"instance_id,omitempty"`
	// Business tracking keyword
	BusinessKey *string `json:"business_key,omitempty"`
	Variables map[string]interface{} `json:"variables,omitempty"`
}

// NewStartInstanceRequest instantiates a new StartInstanceRequest object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewStartInstanceRequest() *StartInstanceRequest {
	this := StartInstanceRequest{}
	return &this
}

// NewStartInstanceRequestWithDefaults instantiates a new StartInstanceRequest object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewStartInstanceRequestWithDefaults() *StartInstanceRequest {
	this := StartInstanceRequest{}
	return &this
}

// GetInstanceId returns the InstanceId field value if set, zero value otherwise.
func (o *StartInstanceRequest) GetInstanceId() string {
	if o == nil || IsNil(o.InstanceId) {
		var ret string
		return ret
	}
	return *o.InstanceId
}

// GetInstanceIdOk returns a tuple with the InstanceId field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *StartInstanceRequest) GetInstanceIdOk() (*string, bool) {
	if o == nil || IsNil(o.InstanceId) {
		return nil, false
	}
	return o.InstanceId, true
}

// HasInstanceId returns a boolean if a field has been set.
func (o *StartInstanceRequest) HasInstanceId() bool {
	if o != nil && !IsNil(o.InstanceId) {
		return true
	}

	return false
}

// SetInstanceId gets a reference to the given string and assigns it to the InstanceId field.
func (o *StartInstanceRequest) SetInstanceId(v string) {
	o.InstanceId = &v
}

// GetBusinessKey returns the BusinessKey field value if set, zero value otherwise.
func (o *StartInstanceRequest) GetBusinessKey() string {
	if o == nil || IsNil(o.BusinessKey) {
		var ret string
		return ret
	}
	return *o.BusinessKey
}

// GetBusinessKeyOk returns a tuple with the BusinessKey field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *StartInstanceRequest) GetBusinessKeyOk() (*string, bool) {
	if o == nil || IsNil(o.BusinessKey) {
		return nil, false
	}
	return o.BusinessKey, true
}

// HasBusinessKey returns a boolean if a field has been set.
func (o *StartInstanceRequest) HasBusinessKey() bool {
	if o != nil && !IsNil(o.BusinessKey) {
		return true
	}

	return false
}

// SetBusinessKey gets a reference to the given string and assigns it to the BusinessKey field.
func (o *StartInstanceRequest) SetBusinessKey(v string) {
	o.BusinessKey = &v
}

// GetVariables returns the Variables field value if set, zero value otherwise.
func (o *StartInstanceRequest) GetVariables() map[string]interface{} {
	if o == nil || IsNil(o.Variables) {
		var ret map[string]interface{}
		return ret
	}
	return o.Variables
}

// GetVariablesOk returns a tuple with the Variables field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *StartInstanceRequest) GetVariablesOk() (map[string]interface{}, bool) {
	if o == nil || IsNil(o.Variables) {
		return map[string]interface{}{}, false
	}
	return o.Variables, true
}

// HasVariables returns a boolean if a field has been set.
func (o *StartInstanceRequest) HasVariables() bool {
	if o != nil && !IsNil(o.Variables) {
		return true
	}

	return false
}

// SetVariables gets a reference to the given map[string]interface{} and assigns it to the Variables field.
func (o *StartInstanceRequest) SetVariables(v map[string]interface{}) {
	o.Variables = v
}

func (o StartInstanceRequest) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o StartInstanceRequest) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	if !IsNil(o.InstanceId) {
		toSerialize["instance_id"] = o.InstanceId
	}
	if !IsNil(o.BusinessKey) {
		toSerialize["business_key"] = o.BusinessKey
	}
	if !IsNil(o.Variables) {
		toSerialize["variables"] = o.Variables
	}
	return toSerialize, nil
}

type NullableStartInstanceRequest struct {
	value *StartInstanceRequest
	isSet bool
}

func (v NullableStartInstanceRequest) Get() *StartInstanceRequest {
	return v.value
}

func (v *NullableStartInstanceRequest) Set(val *StartInstanceRequest) {
	v.value = val
	v.isSet = true
}

func (v NullableStartInstanceRequest) IsSet() bool {
	return v.isSet
}

func (v *NullableStartInstanceRequest) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableStartInstanceRequest(val *StartInstanceRequest) *NullableStartInstanceRequest {
	return &NullableStartInstanceRequest{value: val, isSet: true}
}

func (v NullableStartInstanceRequest) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableStartInstanceRequest) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}