# typescript/src/api/src/models/ProcessInstance.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NativeBPM Engine REST API
 * REST API for managing, executing, and monitoring workflows, human tasks, incidents, and outgoing webhooks inside the NativeBPM Cloud-Native engine.  ### Authentication Requests must include a session cookie or a Bearer API token: `Authorization: Bearer <API_TOKEN>` 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProcessInstance
 */
export interface ProcessInstance {
    /**
     * 
     * @type {string}
     * @memberof ProcessInstance
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessInstance
     */
    processId: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessInstance
     */
    definitionHash: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessInstance
     */
    businessKey: string;
    /**
     * Base64/raw JSON encoded internal Wazero process engine state representation
     * @type {string}
     * @memberof ProcessInstance
     */
    state: string;
    /**
     * 
     * @type {number}
     * @memberof ProcessInstance
     */
    version: number;
    /**
     * 
     * @type {boolean}
     * @memberof ProcessInstance
     */
    completed: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ProcessInstance
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ProcessInstance
     */
    tenantId: string;
}

/**
 * Check if a given object implements the ProcessInstance interface.
 */
export function instanceOfProcessInstance(value: object): value is ProcessInstance {
    if (!('id' in value) || value['id'] === undefined) return false;
    if ((!('processId' in value) && !('process_id' in value)) || (value['processId'] === undefined && value['process_id'] === undefined)) return false;
    if ((!('definitionHash' in value) && !('definition_hash' in value)) || (value['definitionHash'] === undefined && value['definition_hash'] === undefined)) return false;
    if ((!('businessKey' in value) && !('business_key' in value)) || (value['businessKey'] === undefined && value['business_key'] === undefined)) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    if (!('completed' in value) || value['completed'] === undefined) return false;
    if ((!('updatedAt' in value) && !('updated_at' in value)) || (value['updatedAt'] === undefined && value['updated_at'] === undefined)) return false;
    if ((!('tenantId' in value) && !('tenant_id' in value)) || (value['tenantId'] === undefined && value['tenant_id'] === undefined)) return false;
    return true;
}

export function ProcessInstanceFromJSON(json: any): ProcessInstance {
    return ProcessInstanceFromJSONTyped(json, false);
}

export function ProcessInstanceFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProcessInstance {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'processId': json['process_id'],
        'definitionHash': json['definition_hash'],
        'businessKey': json['business_key'],
        'state': json['state'],
        'version': json['version'],
        'completed': json['completed'],
        'updatedAt': (new Date(json['updated_at'])),
        'tenantId': json['tenant_id'],
    };
}

export function ProcessInstanceToJSON(json: any): ProcessInstance {
    return ProcessInstanceToJSONTyped(json, false);
}

export function ProcessInstanceToJSONTyped(value?: ProcessInstance | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'process_id': value['processId'],
        'definition_hash': value['definitionHash'],
        'business_key': value['businessKey'],
        'state': value['state'],
        'version': value['version'],
        'completed': value['completed'],
        'updated_at': value['updatedAt'].toISOString(),
        'tenant_id': value['tenantId'],
    };
}