# typescript/src/api/src/models/StartInstanceRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NativeBPM Engine REST API
 * REST API for managing, executing, and monitoring workflows, human tasks, incidents, and outgoing webhooks inside the NativeBPM Cloud-Native engine.  ### Authentication Requests must include a session cookie or a Bearer API token: `Authorization: Bearer <API_TOKEN>` 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StartInstanceRequest
 */
export interface StartInstanceRequest {
    /**
     * Optional user-generated UUID to enforce idempotency
     * @type {string}
     * @memberof StartInstanceRequest
     */
    instanceId?: string;
    /**
     * Business tracking keyword
     * @type {string}
     * @memberof StartInstanceRequest
     */
    businessKey?: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof StartInstanceRequest
     */
    variables?: { [key: string]: any; };
}

/**
 * Check if a given object implements the StartInstanceRequest interface.
 */
export function instanceOfStartInstanceRequest(value: object): value is StartInstanceRequest {
    return true;
}

export function StartInstanceRequestFromJSON(json: any): StartInstanceRequest {
    return StartInstanceRequestFromJSONTyped(json, false);
}

export function StartInstanceRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): StartInstanceRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'instanceId': json['instance_id'] == null ? undefined : json['instance_id'],
        'businessKey': json['business_key'] == null ? undefined : json['business_key'],
        'variables': json['variables'] == null ? undefined : json['variables'],
    };
}

export function StartInstanceRequestToJSON(json: any): StartInstanceRequest {
    return StartInstanceRequestToJSONTyped(json, false);
}

export function StartInstanceRequestToJSONTyped(value?: StartInstanceRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'instance_id': value['instanceId'],
        'business_key': value['businessKey'],
        'variables': value['variables'],
    };
}